feat(ui): TE-1162 folder restructuring
